Extract `host_udf` header
